Drink equation

+++ app/src/main/java/com/example/elyseturner/howhappycanmyhourbe/activities/models/ExerciseModel.java
@@ -8,7 +8,7 @@ import android.os.Parcelable;
  */
 public class ExerciseModel implements Parcelable {
     private String name;
-    private int calories;
+    private double calories;
 
     public ExerciseModel() {
         this.calories = calories;
